requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
black==26.3.1
celery==5.6.2
certifi==2026.2.25
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.12
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-tenants==3.10.1
django-timezone-field==7.2.1
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
idna==3.11
isort==8.0.1
kombu==5.6.2
mccabe==0.7.0
midtransclient==1.4.2
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
qrcode==8.2
redis==7.3.0
reportlab==4.4.10
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
xendit-python==7.0.0
